Refactor: #120 TodoSchedulerService 제거 및 TodoService로 이동 - 다른 서비스로 뺄 만큼 메서드가 많지 않기 때문에 코드 이동
